ADD MULTIPLEXING SERVER USING SELECT():
Server accepts simultaneous connexion and does not crash when one client
disconnect.

Misunderstanding on my part : when select returns a positive fd > 0, it
means there is data to be read. However, this data can be from
the listening socket or from an existing connection but either or, both
fd will be witin the workingSet.
If the data comes from the listening socket, then a new fd is created
and is added to the workingSet.
If it's not, then the data comes from an established connection and can
be read as such.

Author:    Elena-Lou <elouisiade@live.fr>
Date:      Wed May 17 19:03:36 2023 +0200