ADD CLASS SERVER:
Adding class server.
This creates the socket at instantiation.
The class sould be instantiated in a try catch structure.

main is left unmodified for now.
providing a main_classified.cpp to test the code, to be deleted
obviously.

Makefile is modified accordingly (left on main.cpp, no rules for
main_classified was left in the Makefile for consistency).