UPDATE NEW/READ/WRITE LOOPS (MAJOR):
This is a major commit with huge modifications to previous work done.

Event loop is now mostly embedded in methods from the Server class.
This loop now works as follows :
- check for new clients.
- check for clients to read from.
- check for clients to write to.

Clients are now stored in a set of Client* in the Server class.
Disconnected/timed-out/kicked clients are removed from this set at each
step of the event loop.

The loop no longer loops on fds but loops on logged-in clients instead.
No longer looping for nothing on fds no longer in use.

Buffers still need to be worked on.
Write loop is a copy of the existing read loop with just a flag added
to avoid some errors later on.
The overall system of buffer needs to be improved and used more
efficiently.

Buffers are stored in the client instances for now, using a define
IRC_BUFFER_SIZE.

Error messages are mostly printing on cerr instead of cout while cout is
still in use at this point in the project for useful dev information
that will be removed later on.

Changes to be committed:
	modified:   Client.cpp
	modified:   Client.hpp
	modified:   Server.cpp
	modified:   Server.hpp
	modified:   irc.hpp
	modified:   main.cpp

	deleted:    main_classified.cpp (no longer needed for
	demonstration purpose as main was updated)