ADD CLASS CHANNEL:
Adding class channel.
Preparing a lot of function that will be modified later to suit the
needs of the project.

The Channels are stored as an attribute of the server class, in a
container.
The Client(s) keep track of which Channel(s) they are in.
The Channel(s) keep track of which Client(s) are logged in.

Changes to be committed:
	new file:   Channel.cpp
	new file:   Channel.hpp
	modified:   Client.cpp
	modified:   Client.hpp
	modified:   Makefile
	modified:   Server.cpp
	modified:   Server.hpp
	modified:   irc.hpp