UPDATE ALLOCATION IN CONTAINERS:
Now using a list of Channel(s), a map of Client(s).
Inside the Client(s) storing a set of Channel(s) addresses.
Inside the Channel(s) storing a map of client addresses.

No longer closing the fds in the Client destructor.

For future references:
	the server connects (constructors) and disconnets users.
	the channel adds and removes users (clients).
	the users join and quit channels.

The channel keeps tracks of how much users are in it.
The server checks if the channels must be deleted (no users).
The client takes care of itself only.