FIX NICK:
Nick now checks for previously used nick as expected.
Tested error messages.
Also tested for broadcast msg but it seems like irssi is blocking them as
long as user is not in a channel.

Weird behaviour on first call of NICK when someone else already owns
the nick...

Need to investigate that further.
Dealt with already taken name appending a '_' to it untill the name is
not taken.

This has no size limitation so this could be considered risky.