FIX NICK ERRMSG && ADD RPL ADDITIONAL SOURCE:
Fixing the nick error messages.

The way the RPL are built is basically centered around a ":"

We have
<source> <optional-params> ":" <msg && other args>

The source is automaticaly recovered in the send RPL method.
However the optional params were not dealt with and are now added
in an overloaded method.

In case RPL need to be sent to several users methods for Server and
Channel will be added later.