ADD PASS REWORK RPL PRINTER:
Rebuilt the way the RPLs are printed to remove those ugly _S defines.

Now using a setw() and setfill() call on the message builder to
display RPL using int(s) only.

Updated the error() of :
- join
- kick
- nick

Added Pass untested yet.

Kick does not kick yet and only checks simple errors.
Still a lot of work to be done on it.