FIX CONSTRUCTOR, CMD PARSING (SERVER):
In operator= a lot of attributes were missing. Not really important
as no copies of the server are made but still.

Constructor now takes a string instead of a const char * as parameter.
This was misleading and caused errors in the ways the server decided
if it was protected by password.

In parsing command, fixed the index set to 9 as default resulting in CAP
being triggered if no command were found.
I did not realise it was set to max value as setting it to -1 is
preferable in order to not have to increase it each time we add a new
command.

Also fixed an error in parsing command that led to the input being
left empty if no parameters were given.
This error was not spotted due to using only irssi for the tests.

Updates to check for ERR_NEEDMOREPARAMS with all the already implemented
params are likely to come in a near future.