FIX SEGFAULT AT CHANNEL REMOVAL:
Removing the last element of a Deque compromises the iterators.
Therefore when cycling with a loop of iterator removing iterators
on a deque, we need to check that it's size is not zero at each loop
turn.

This is slowing down the process, but deque is normaly not intended
to such use whereas list are good at it.

For reasons that we will not reexplicit here we chose to use a deque
instead of a list knowing that.