FIX PRIVMSG DUPLICATE MESSAGE:
The way the message was parsed and rebuilt in privmsg was biased and led
to duplicate messages.

Because the message needs to start with a ":" in irssi we are forcing
this behaviour.

Also no longer sending messages to the author of the message unless
the author is also the dest.