UPDATE COMPLETE REGISTRATION:
Now sending RPL_LUSERCLIENT and RPL_LUSERME and more
when first registering to the server.

Adding getters for Server for the nb of connected clients, Ops and
channels.

Adding a map of Ops in the server.
Adding a password for Ops in irc.hpp
Should update command OPER accordingly in the next commit